list route security added

+++ controllers/api/list-routes.js
@@ -3,6 +3,8 @@ const { User, List, Item, Group, GroupList } = require("../../models");
 
 // api/lists endpoint
 
+// Get on all lists disabled. Only enable for testing purposes.
+/*
 router.get("/", async (req, res) => {
   try {
     const listData = await List.findAll({
@@ -17,7 +19,7 @@ router.get("/", async (req, res) => {
   } catch (err) {
     res.status(500).json(err);
   }
-});
+});*/
 
 router.get("/:id", async (req, res) => {
   try {
@@ -31,6 +33,10 @@ router.get("/:id", async (req, res) => {
 
     if (!listData) {
       res.status(404).json({ message: "No list with that ID." });
+      return;
+    }
+    if (listData.user_id !== req.session.userID) {
+      res.status(401).json({message: "This is not your list. Please log in as the correct user."});
       return;
     }
     res.status(200).json(listData);
@@ -49,18 +55,21 @@ router.post("/", async (req, res) => {
     }
     */
   List.create(req.body)
-    .then((list) => {
-      if (req.body.groupIds) {
-        const groupListIdArr = req.body.groupIds.map((group_id) => {
-          return {
-            list_id: list.id,
-            group_id,
-          };
-        });
-        GroupList.bulkCreate(groupListIdArr);
-      }
-      return true;
-    })
+    // initializing a list with GroupLists is disabled to protect existing groups. Enable only for testing.
+    /*
+      .then((list) => {
+        if (req.body.groupIds) {
+          const groupListIdArr = req.body.groupIds.map((group_id) => {
+            return {
+              list_id: list.id,
+              group_id,
+            };
+          });
+          GroupList.bulkCreate(groupListIdArr);
+        }
+        return true;
+      })
+    */
     .then((results) => res.status(200).json(results))
     .catch((err) => {
       console.log(err);
@@ -68,9 +77,16 @@ router.post("/", async (req, res) => {
     });
 });
 
-//Alter a group. Can change group title and remove a grouplist if it's included.
+// Put on a list.
+// Can remove from a group, and change the list title.
 router.put("/:id", async (req, res) => {
   try {
+    const thisList = await List.findByPk(req.params.id);
+    if(thisList.user_id !== req.session.userID) {
+      res.status(401).json({message: "This is not your list. Please log in as the correct user."});
+      return;
+    }
+
     if (req.body.removedGroup) {
       await GroupList.destroy({
         where: { group_id: req.body.removedGroup, list_id: req.params.id },
@@ -83,11 +99,7 @@ router.put("/:id", async (req, res) => {
       );
       res.status(200).json(listData);
     } else
-      res
-        .status(200)
-        .json({
-          message: `group with ID ${req.body.removedGroup} removed from this list.`,
-        });
+      res.status(200).json({message: `group with ID ${req.body.removedGroup} removed from this list.`});
   } catch (err) {
     res.status(500).json(err);
   }
@@ -96,11 +108,18 @@ router.put("/:id", async (req, res) => {
 //delete to destroy a list
 router.delete("/:id", async (req, res) => {
   try {
+    const thisList = await List.findByPk(req.params.id);
+
+    if(thisList.user_id !== req.session.userID) {
+      res.status(401).json({message: "This is not your list. Please log in as the correct user."});
+      return;
+    }
+
     await Item.destroy({
-      where: {list_id: req.params.id},
+      where: { list_id: req.params.id },
     });
     await GroupList.destroy({
-      where: {list_id: req.params.id},
+      where: { list_id: req.params.id },
     });
     const listData = await List.destroy({
       where: { id: req.params.id },
